+++ src/app/components/dummy-fb-metadata-modal.js
@@ -4,8 +4,11 @@ import { inject as service } from '@ember/service';
 
 export default Component.extend({
   mockBackend: service(),
+  navigationState: service(),
 
   file: reads('mockBackend.entityRecords.file.firstObject'),
 
   open: true,
+
+  previewMode: reads('navigationState.aspectOptions.previewMode'),
 });