+++ src/app/locales/en/components/edit-permissions-modal.js
@@ -1,6 +1,10 @@
 export default {
-  header: 'Edit permissions',
+  header: {
+    edit: 'Edit permissions',
+    show: 'Permissions',
+  },
   cancel: 'Cancel',
+  close: 'Close',
   save: 'Save',
   modifyingPermissions: 'modifying permissions',
   permissionsModifySuccess: 'Permissions has been modified successfully',
@@ -11,8 +15,15 @@ export default {
     acl: 'ACL',
   },
   editAnyway: 'Edit anyway',
-  differentPermissionsTypes: 'Selected files use different permissions types. You can choose one of the types (POSIX or ACL) to be applied for all files - current permissions will be <strong>irreversibly overwritten</strong> upon save.',
-  differentPosix: 'Selected files have different POSIX permissions. You can edit them, but saving your changes will cause the permissions of all files to be <strong>irreversibly overwritten</strong> with the new ones.',
+  previewModeAcl: 'Selected file(s) use Access Control List permissions, whose details cannot be viewed in share browser.',
+  differentPermissionsTypes: {
+    edit: 'Selected files use different permissions types. You can choose one of the types (POSIX or ACL) to be applied for all files - current permissions will be <strong>irreversibly overwritten</strong> upon save.',
+    show: 'Selected files use different permissions types (POSIX or ACL). Only POSIX permissions can be viewed in share browser.',
+  },
+  differentPosix: {
+    edit: 'Selected files have different POSIX permissions. You can edit them, but saving your changes will cause the permissions of all files to be <strong>irreversibly overwritten</strong> with the new ones.',
+    show: 'Selected files have different POSIX permissions.',
+  },
   differentAcl: 'Selected files have different ACL rules. You can edit them, but saving your changes will cause the ACL rules of all files to be <strong>irreversibly overwritten</strong> with the new ones.',
   noAclDueToMixedFileTypes: 'Cannot view ACL of files and directories at the same time. Select only files or only directories to enable ACL editor.',
   ownerSystemSubject: 'owner',